Add files via upload
The main objectives of the project were to develop an algorithm that  will be used to identify answers related to user submitted questions.  To develop a database were all the related data will be stored and to  develop a web interface. The web interface developed had two  
parts, one for simple users and one for the administrator. A  background research took place, which included an overview of the  conversation procedure and any relevant chat bots available. A  database was developed, which stores information about questions,  answers, keywords, logs and feedback messages.